refactor(feedback-widget): extract PinMarker and PinActionCluster
Move the two leaf pin components out of the orchestrator into
src/components/FeedbackWidget/pin/. PinMarker is purely visual;
PinActionCluster owns its own internal menuOpen state. Drop
AVATAR_COLORS and NOISE_OVERLAY_BG from the orchestrator's import
list (now only used inside the extracted modules).

Adds unit tests for both components covering the outline branch on
PinMarker, the resolved/unresolved branches on PinActionCluster's
Approve button, the Reopen/Edit/Delete kebab menu paths, and the
backdrop click that closes the menu.

Adds an integration test for clicking the comment popover scrim to
return to selecting mode — exercises clearImage() in the cancel path
which was not covered today.
